Fix tests, create module for reload utils
